+++ rf4s/component/friction_brake.py
@@ -19,6 +19,7 @@ MIN_FRICTION_BRAKE = 0
 UP = 1
 DOWN = -1
 FRICTION_BRAKE_MONITOR_DELAY = 2
+FRICTION_BRAKE_CHANGE_DELAY = 0.04
 
 
 class FrictionBrake:
@@ -56,20 +57,30 @@ class FrictionBrake:
         self.cur = Value("i", self.fb_cfg.INITIAL)
         self.monitor_process = Process(target=monitor_friction_brake, args=(self,))
 
-    def reset(self, target: int) -> None:
+    def reset(self, target: int, to_max: bool=False) -> None:
         """Reset the friction brake to the target value.
 
         :param target: Target friction brake value.
         :type target: int
+        :param to_max: Increase to 30 before setting it to target, defaults to False.
+        :type to_max: bool, optional
         """
         logger.info("Resetting friction brake")
-        for _ in range(MAX_FRICTION_BRAKE):
-            pag.scroll(UP)
-
-        diff = MAX_FRICTION_BRAKE - target
-        for _ in range(abs(diff)):
-            pag.scroll(DOWN)
-        self.cur.value = target
+        with self.lock:
+            sleep(FRICTION_BRAKE_CHANGE_DELAY) # might be called after change()
+            if to_max:
+                for _ in range(MAX_FRICTION_BRAKE):
+                    pag.scroll(UP, _pause=False)
+                    sleep(FRICTION_BRAKE_CHANGE_DELAY)
+                cur = MAX_FRICTION_BRAKE
+            else:
+                cur = self.cur.value
+            diff = cur - target
+            print(cur, target, diff)
+            for _ in range(abs(diff)):
+                pag.scroll(DOWN, _pause=False)
+                sleep(FRICTION_BRAKE_CHANGE_DELAY)
+            self.cur.value = target
 
     def change(self, increase: bool) -> None:
         """Increase or decrease the friction brake.